requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-allauth==65.4.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
psycopg2-binary==2.9.10
Pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.0
dj-database-url==2.2.0 